requirements.txt
opencv-python
yt-dlp
fpdf2
easyocr
openai
python-dotenv
google-generativeai
numpy
